+++ doc/code/gbs/graph.similarity.rst
